+++ Example4.MvcWebApp.IndividualAccounts/Models/AuthUserDisplay.cs
@@ -11,15 +11,16 @@ namespace Example4.MvcWebApp.IndividualAccounts.Models
     public class AuthUserDisplay
     {
         [MaxLength(AuthDbConstants.UserNameSize)]
-        public string UserName { get;  set; }
+        public string UserName { get; private set; }
         [Required(AllowEmptyStrings = false)]
         [MaxLength(AuthDbConstants.EmailSize)]
-        public string Email { get; set; }
+        public string Email { get; private set; }
         [Required(AllowEmptyStrings = false)]
         [MaxLength(AuthDbConstants.UserIdSize)]
-        public string UserId { get; set; }
-        public int NumRoles { get; set; }
-        public bool HasTenant { get; set; }
+        public string UserId { get; private set; }
+        public string[] RoleNames { get; private set; }
+        public bool HasTenant => TenantName != null;
+        public string TenantName { get; private set; }
 
         public static IQueryable<AuthUserDisplay> SelectQuery(IQueryable<AuthUser> inQuery)
         {
@@ -28,19 +29,24 @@ namespace Example4.MvcWebApp.IndividualAccounts.Models
                 UserName = x.UserName,
                 Email = x.Email,
                 UserId = x.UserId,
-                NumRoles = x.UserRoles.Count(),
-                HasTenant = x.TenantId != null
+                RoleNames = x.UserRoles.Select(y => y.RoleName).ToArray(),
+                TenantName = x.UserTenant.TenantName
             });
         }
 
         public static AuthUserDisplay DisplayUserInfo(AuthUser authUser)
         {
-            return new AuthUserDisplay
+            var result = new AuthUserDisplay
             {
                 UserName = authUser.UserName,
                 Email = authUser.Email,
                 UserId = authUser.UserId,
+                TenantName = authUser.UserTenant?.TenantName
             };
+            if (authUser.UserRoles != null)
+                result.RoleNames = authUser.UserRoles.Select(y => y.RoleName).ToArray();
+
+            return result;
         }
     }
 }