+++ Example4.ShopServiceLayer/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Example4.ShopServiceLayer
+{
+    public class Class1
+    {
+    }
+}